Keep the CPU pool fractional, at the top as well as the bottom
Section 1.2 says an activation costs cycles "in fractions", and thirteen items
in the catalogue declare a fractional cost. The backlog said those costs were
floored to a whole cycle. They are not, any more - the max(1, ...) it names is
gone, and a 0.3 cost now takes 0.3 - so that entry goes with this.

What was still whole was the pool. `max_cpu` was an int while the level was a
float, and min() returns whichever operand it picked, so every time
regeneration topped the pool up the level came back an int: a full pool
reported 3 and a spent one 2.7, out of the same field. Infrastructure adds to
that pool, and nothing says those additions are whole either.

Every get_cpu_cost said it returned an int while returning the float its own
field holds. Annotations only, but eight of them, all disagreeing with the
thing they annotate.

The test covering section 1.2 had comments saying the pool starts at ten and
regenerates two a second, directly above assertions checking three and one.
Ten is also exactly what the client drew for every battle until last commit,
so that comment may be where it came from.

+++ server/battle_engine.py
@@ -150,9 +150,13 @@ class Player:
     quota: int  # Current
     max_quota: int  # Based on round
 
-    # Section 1.2: CPU Cycles (Stamina)
+    # Section 1.2: CPU Cycles (Stamina). All three are fractional, and the
+    # pool has to be as much as the level does: min() returns whichever
+    # operand it picked, so an int pool turned a full float pool back into an
+    # int every time regeneration topped it up. Infrastructure adds to it, and
+    # nothing says those additions are whole either.
     cpu: float  # Current cycles
-    max_cpu: int = 3
+    max_cpu: float = 3.0
     cpu_regen: float = 1.0  # Per second
 
     # Section 3: Buffs & Debuffs
@@ -263,8 +267,8 @@ class BattleSimulator:
         # from max_cpu, which is what happened when the pool was last changed.
         player1 = Player(id=1, quota=quota, max_quota=quota, cpu=0.0)
         player2 = Player(id=2, quota=quota, max_quota=quota, cpu=0.0)
-        player1.cpu = float(player1.max_cpu)
-        player2.cpu = float(player2.max_cpu)
+        player1.cpu = player1.max_cpu
+        player2.cpu = player2.max_cpu
 
         # Deep copy items to avoid mutation
         p1_items = deepcopy(p1_items)
